refactor: remove mixin, move position property to ContextMenu
